+++ workers/go/kitchensink/kitchen_sink.go
@@ -1,4 +1,4 @@
-package workflows
+package kitchensink
 
 import (
 	"encoding/json"
