+++ workers/go/worker/worker.go
@@ -3,10 +3,13 @@ package worker
 import (
 	"fmt"
 
+	"github.com/temporalio/omes/workers/go/activities"
+	"github.com/temporalio/omes/workers/go/kitchensink"
+	"github.com/temporalio/omes/workers/go/throughputstress"
+	"go.temporal.io/sdk/activity"
+
 	"github.com/spf13/cobra"
 	"github.com/temporalio/omes/cmd/cmdoptions"
-	"github.com/temporalio/omes/workers/go/activities"
-	"github.com/temporalio/omes/workers/go/workflows"
 	"go.temporal.io/sdk/client"
 	"go.temporal.io/sdk/worker"
 	"go.temporal.io/sdk/workflow"
@@ -52,17 +55,18 @@ func (a *App) Run(cmd *cobra.Command, args []string) {
 
 func runWorkers(client client.Client, taskQueues []string) error {
 	errCh := make(chan error, len(taskQueues))
-	activityInstance := activities.Activities{
+	tpsActivities := throughputstress.Activities{
 		Client: client,
 	}
 	for _, taskQueue := range taskQueues {
 		taskQueue := taskQueue
 		go func() {
 			w := worker.New(client, taskQueue, worker.Options{})
-			w.RegisterWorkflowWithOptions(workflows.KitchenSinkWorkflow, workflow.RegisterOptions{Name: "kitchenSink"})
-			w.RegisterWorkflowWithOptions(workflows.ThroughputStressWorkflow, workflow.RegisterOptions{Name: "throughputStress"})
-			w.RegisterWorkflowWithOptions(workflows.ThroughputStressChild, workflow.RegisterOptions{Name: "throughputStressChild"})
-			w.RegisterActivity(&activityInstance)
+			w.RegisterWorkflowWithOptions(kitchensink.KitchenSinkWorkflow, workflow.RegisterOptions{Name: "kitchenSink"})
+			w.RegisterWorkflowWithOptions(throughputstress.ThroughputStressWorkflow, workflow.RegisterOptions{Name: "throughputStress"})
+			w.RegisterWorkflowWithOptions(throughputstress.ThroughputStressChild, workflow.RegisterOptions{Name: "throughputStressChild"})
+			w.RegisterActivity(&tpsActivities)
+			w.RegisterActivityWithOptions(activities.Noop, activity.RegisterOptions{Name: "noop"})
 			errCh <- w.Run(worker.InterruptCh())
 		}()
 	}
