+++ workers/python/activities.py
@@ -1,6 +1,6 @@
 from temporalio import activity
 
 
-@activity.defn(name="Noop")
+@activity.defn(name="noop")
 async def noop_activity():
     return
